Make Glide HUD adaptable for more elements
