+++ src/main/resources/assets/lemclienthelper/textures/gui/speedometer/arrow/0.png
