+++ src/main/resources/assets/lemclienthelper/textures/gui/speedometer/arrow/1.png
