+++ src/main/resources/assets/lemclienthelper/textures/gui/speedometer/arrow/10.png
