+++ src/main/resources/assets/lemclienthelper/textures/gui/speedometer/arrow/11.png
