+++ src/main/resources/assets/lemclienthelper/textures/gui/speedometer/arrow/2.png
