+++ src/main/resources/assets/lemclienthelper/textures/gui/speedometer/arrow/3.png
