+++ src/main/resources/assets/lemclienthelper/textures/gui/speedometer/arrow/4.png
