+++ src/main/resources/assets/lemclienthelper/textures/gui/speedometer/arrow/5.png
