+++ src/main/resources/assets/lemclienthelper/textures/gui/speedometer/arrow/6.png
