+++ src/main/resources/assets/lemclienthelper/textures/gui/speedometer/arrow/7.png
