+++ src/main/resources/assets/lemclienthelper/textures/gui/speedometer/arrow/8.png
