+++ src/main/resources/assets/lemclienthelper/textures/gui/speedometer/arrow/9.png
