+++ src/main/resources/assets/lemclienthelper/textures/gui/speedometer/speedometer.png
